requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
asyncio==3.4.3
attrs==23.2.0
av==12.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
browser-cookie3==0.19.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
crypto==1.4.1
cryptography==43.0.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==4.4.2
defusedxml==0.7.1
discord==2.3.2
discord-protos==0.0.2
discord-webhook==1.3.1
discord.py==2.4.0
discord.py-self @ git+https://github.com/dolfies/discord.py-self@38780526e9db5c2e14be8f220cb36b5e3a847c6f
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gast==0.6.0
google-crc32c==1.6.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.2
multidict==6.0.5
Naked==0.1.32
namex==0.0.8
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.1
numpy==1.26.4
onnxruntime-directml==1.17.3
openai==0.27.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2024.1.0
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pylibsrtp==0.10.0
PyMsgBox==1.0.9
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytweening==1.2.0
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scipy==1.14.1
selenium==4.25.0
setuptools==70.3.0
shellescape==3.8.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
tinycss2==1.3.0
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.7
vdf==3.4
virtualenv==20.27.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
websockify==0.12.0
Werkzeug==3.0.3
wheel==0.44.0
win32-setctime==1.1.0
windows-capture==1.3.6
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
